Rename C files to avoid .o name clashes on OSX
