+++ cfov.c
@@ -11,7 +11,7 @@
 #include <math.h>
 #include <float.h>
 #include <assert.h>
-#include "fov.h"
+#include "cfov.h"
 
 /*
 +---++---++---++---+
